@@ -1,12 +1,16 @@
 module github.com/gamebtc/devicedetector
 
-go 1.13
+go 1.22
 
 require (
 	github.com/dlclark/regexp2 v1.2.0
-	github.com/google/go-cmp v0.3.1
 	github.com/mcuadros/go-version v0.0.0-20190830083331-035f6764e8d2
-	github.com/pkg/errors v0.8.1 // indirect
+	github.com/stretchr/testify v1.9.0
 	gopkg.in/yaml.v2 v2.3.0
-	gotest.tools v2.2.0+incompatible
+)
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	gopkg.in/yaml.v3 v3.0.1 // indirect
 )
