@@ -1,9 +1,10 @@
 package parser
 
 import (
-	"gotest.tools/assert"
 	"path/filepath"
 	"testing"
+
+	"github.com/stretchr/testify/require"
 )
 
 var botParser = NewBot(filepath.Join(dir, FixtureFileBot))
@@ -21,19 +22,19 @@ func TestGetInfoFromUABot(t *testing.T) {
 			Url:  "http://www.google.com",
 		},
 	}
-	assert.DeepEqual(t, expected, info)
+	require.EqualValues(t, expected, info)
 }
 
 func TestParseNoDetails(t *testing.T) {
 	ua := `Googlebot/2.1 (http://www.googlebot.com/bot.html)`
 	botParser.DiscardDetails(true)
 	info := botParser.Parse(ua)
-	assert.Check(t, info != nil)
+	require.NotNil(t, info)
 }
 
 func TestParseNoBot(t *testing.T) {
 	ua := `Mozilla/4.0 (compatible; MSIE 9.0; Windows NT 6.1; SV1; SE 2.x)`
 	botParser.DiscardDetails(false)
 	info := botParser.Parse(ua)
-	assert.Check(t, info == nil)
+	require.Nil(t, info)
 }
