@@ -4,7 +4,7 @@ import (
 	"path/filepath"
 	"testing"
 
-	"gotest.tools/assert"
+	"github.com/stretchr/testify/require"
 )
 
 func TestOsParse(t *testing.T) {
@@ -27,12 +27,12 @@ func TestOsParse(t *testing.T) {
 		//assert.Check(t, NameEqual(item.ShortName, r.ShortName), )
 		//assert.Check(t, NameEqual(item.Version, r.Version), )
 		//assert.Check(t, NameEqual(item.Platform, r.Platform), )
-		assert.DeepEqual(t, &item.OsMatchResult, r)
+		require.EqualValues(t, &item.OsMatchResult, r)
 	}
 }
 
 func TestOsInGroup(t *testing.T) {
-	for os, _ := range OperatingSystems {
+	for os := range OperatingSystems {
 		contains := false
 		for _, familyOs := range OsFamilies {
 			if ArrayContainsString(familyOs, os) {
@@ -40,20 +40,20 @@ func TestOsInGroup(t *testing.T) {
 				break
 			}
 		}
-		assert.Check(t, contains)
+		require.True(t, contains)
 	}
 }
 
 func TestGetNameFromId(t *testing.T) {
 	data := [][]string{
-		[]string{"DEB", "4.5", "Debian 4.5"},
-		[]string{"WRT", "", "Windows RT"},
-		[]string{"WIN", "98", "Windows 98"},
-		[]string{"XXX", "4.5", ""},
+		{"DEB", "4.5", "Debian 4.5"},
+		{"WRT", "", "Windows RT"},
+		{"WIN", "98", "Windows 98"},
+		{"XXX", "4.5", ""},
 	}
 	for _, item := range data {
 		os, version, expected := item[0], item[1], item[2]
 		r := GetOsNameFromId(os, version)
-		assert.Equal(t, r, expected)
+		require.Equal(t, r, expected)
 	}
 }
